requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
spotipy==2.23.0
yt-dlp==2024.11.4
aiofiles==23.2.1
aiohttp==3.9.1
